refactor: folder structure and admission entrypoint, docker and deploying to be done
