+++ src/controller_examples/v_replica_set_controller/exec/mod.rs
@@ -1,4 +1,4 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 pub mod reconciler;
-pub mod source;
+pub mod validator;
