--- src/controller_examples/v_replica_set_controller/exec/source/mod.rs
@@ -1,5 +1,0 @@
-pub mod validator;
-pub mod handler;
-
-pub use validator::*;
-pub use handler::*;
