+++ src/deps_hack/Cargo.toml
@@ -34,4 +34,4 @@ rand = "0.8"
 zookeeper = "0.8"
 chrono = "0.4.19"
 proptest = "1.4.0"
-warp = "0.3"
+warp = { version = "0.3", features = ["tls"] }
